requirements.txt
yt-dlp 
openai-whisper 
torch 
moviepy 
streamlit
-e .
